pyproject.toml
[tool.poetry]
name = "event manager discord bot"
version = "0.1.0"
description = "A Discord Bot built to manage your events"
authors = ["Calphast"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.2"
python-dotenv = "^1.0.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.20"}
aiosqlite = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
